+++ collector/__init__.py
@@ -1,0 +1,1 @@
+"""Coletor automático de fontes e questões do ConcursoAI."""